+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -35,6 +35,7 @@ Other breaking changes:
 * Removed `Snaks` interface, use `SnakList` instead
 * Removed previously deprecated `Entity::getAllSnaks`, use `StatementList::getAllSnaks` instead
 * Removed previously deprecated `EntityId::getPrefixedId`, use `EntityId::getSerialization` instead
+* Removed previously deprecated `Property::newEmpty`, use `Property::newFromType` or `new Property()` instead
 * `Reference` and `ReferenceList`s no longer can not be instantiated with `null`
 * Added `setId` method to `EntityDocument`